refactor: 대기열 토큰 DTO 패키지 이전
